requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
langgraph>=0.2.50
langchain-core>=0.3.17
jinja2>=3.1.4
pydantic>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
mcp>=1.0.0
